Refactor remaining commands
