+++ src/commands/logout/apply-logout.ts
@@ -1,0 +1,8 @@
+import { updateSetting } from '../../utils/settings'
+
+export function applyLogout() {
+  updateSetting('apiToken', null)
+  updateSetting('apiBaseUrl', null)
+  updateSetting('apiProxy', null)
+  updateSetting('enforcedOrgs', null)
+}
