+++ src/commands/organization/get-organizations.ts
@@ -1,46 +1,12 @@
-import meow from 'meow'
 import colors from 'yoctocolors-cjs'
 
 import { Spinner } from '@socketsecurity/registry/lib/spinner'
 
-import { handleApiCall, handleUnsuccessfulApiResponse } from '../utils/api'
-import { AuthError } from '../utils/errors'
-import { getDefaultToken, setupSdk } from '../utils/sdk'
+import { handleApiCall, handleUnsuccessfulApiResponse } from '../../utils/api'
+import { AuthError } from '../../utils/errors'
+import { getDefaultToken, setupSdk } from '../../utils/sdk'
 
-import type { CliSubcommand } from '../utils/meow-with-subcommands'
-
-const description = 'List organizations associated with the API key used'
-
-export const organizationCommand: CliSubcommand = {
-  description,
-  async run(argv, importMeta, { parentName }) {
-    setupCommand(`${parentName} organizations`, description, argv, importMeta)
-    await fetchOrganizations()
-  }
-}
-
-// Internal functions
-
-function setupCommand(
-  name: string,
-  description: string,
-  argv: readonly string[],
-  importMeta: ImportMeta
-) {
-  meow(
-    `
-    Usage
-      $ ${name}
-  `,
-    {
-      argv,
-      description,
-      importMeta
-    }
-  )
-}
-
-async function fetchOrganizations(): Promise<void> {
+export async function getOrganizations(): Promise<void> {
   const apiToken = getDefaultToken()
   if (!apiToken) {
     throw new AuthError(
