+++ src/commands/raw-npx/cmd-raw-npx.ts
@@ -1,0 +1,51 @@
+import meowOrExit from 'meow'
+
+import { runRawNpx } from './run-raw-npx.ts'
+import constants from '../../constants'
+import { commonFlags, validationFlags } from '../../flags'
+import { getFlagListOutput } from '../../utils/output-formatting'
+
+import type { CliCommandConfig } from '../../utils/meow-with-subcommands.ts'
+
+const { NPX } = constants
+
+const config: CliCommandConfig = {
+  commandName: 'raw-npx',
+  description: `Temporarily disable the Socket ${NPX} wrapper`,
+  hidden: false,
+  flags: {
+    ...commonFlags,
+    ...validationFlags
+  },
+  help: (parentName, config) => `
+    Usage
+      $ ${parentName} ${config.commandName} <command>
+
+    Options
+      ${getFlagListOutput(config.flags, 6)}
+
+    Examples
+      $ ${parentName} ${config.commandName} install
+  `
+}
+
+export const cmdRawNpx = {
+  description: config.description,
+  hidden: config.hidden,
+  run
+}
+
+async function run(
+  argv: ReadonlyArray<string>,
+  importMeta: ImportMeta,
+  { parentName }: { parentName: string }
+): Promise<void> {
+  meowOrExit(config.help(parentName, config), {
+    argv,
+    description: config.description,
+    importMeta,
+    flags: config.flags
+  })
+
+  await runRawNpx(argv)
+}
